Se corrigio errores de compatibilidad de tipo de datos en las funciones

+++ src/features/fspectrogram.py
@@ -3,23 +3,25 @@ from matplotlib import pyplot as plt
 import scipy.io.wavfile as wav
 from numpy.lib import stride_tricks
 
+
 """ short time fourier transform of audio signal """
 def stft(sig, frameSize, overlapFac=0.5, window=np.hanning):
     win = window(frameSize)
     hopSize = int(frameSize - np.floor(overlapFac * frameSize))
     
     # zeros at beginning (thus center of 1st window should be for sample nr. 0)
-    samples = np.append(np.zeros(np.floor(frameSize/2.0)), sig)    
+    samples = np.append(np.zeros(np.int(np.floor(frameSize/2.0))), sig)    
     # cols for windowing
     cols = np.ceil( (len(samples) - frameSize) / float(hopSize)) + 1
     # zeros at end (thus samples can be fully covered by frames)
     samples = np.append(samples, np.zeros(frameSize))
     
-    frames = stride_tricks.as_strided(samples, shape=(cols, frameSize), strides=(samples.strides[0]*hopSize, samples.strides[0])).copy()
+    frames = stride_tricks.as_strided(samples, shape=(int(cols), frameSize), strides=(samples.strides[0]*hopSize, samples.strides[0])).copy()
     frames *= win
     
-    return np.fft.rfft(frames)    
-    
+    return np.fft.rfft(frames)  
+
+
 """ scale frequency axis logarithmically """    
 def logscale_spec(spec, sr=44100, factor=20.):
     timebins, freqbins = np.shape(spec)
@@ -27,6 +29,7 @@ def logscale_spec(spec, sr=44100, factor=20.):
     scale = np.linspace(0, 1, freqbins) ** factor
     scale *= (freqbins-1)/max(scale)
     scale = np.unique(np.round(scale))
+    scale = scale.astype(int)
     
     # create spectrogram with new freq bins
     newspec = np.complex128(np.zeros([timebins, len(scale)]))
@@ -46,6 +49,7 @@ def logscale_spec(spec, sr=44100, factor=20.):
             freqs += [np.mean(allfreqs[scale[i]:scale[i+1]])]
     
     return newspec, freqs
+
 
 """ plot spectrogram"""
 def plotstft(audiopath, binsize=2**10, plotpath=None, colormap="jet"):
